Change relative path of style.css and script.js to use "./" instead of "/"
This commit changes the relative path of two files, style.css and script.js, in the index.html file. Instead of using "/style.css" and "/pages/script.js", it now uses "./style.css" and "./pages/script.js", respectively. This change ensures that the files are looked for in the current directory instead of the root directory.